+++ cmd/rqlited/main.go
@@ -305,7 +305,6 @@ func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
 	})
 
 	// Set optional parameters on store.
-	str.StartupOnDisk = cfg.OnDiskStartup
 	str.SetRequestCompression(cfg.CompressionBatch, cfg.CompressionSize)
 	str.RaftLogLevel = cfg.RaftLogLevel
 	str.NoFreeListSync = cfg.RaftNoFreelistSync