+++ http/service.go
@@ -394,7 +394,7 @@ func (s *Service) Start() error {
 	go func() {
 		err := s.httpServer.Serve(s.ln)
 		if err != nil {
-			s.logger.Println("HTTP service stopped:", err.Error())
+			s.logger.Printf("HTTP service on %s stopped: %s", s.ln.Addr().String(), err.Error())
 		}
 	}()
 	s.logger.Println("service listening on", s.Addr())
@@ -404,6 +404,7 @@ func (s *Service) Start() error {
 
 // Close closes the service.
 func (s *Service) Close() {
+	s.logger.Println("closing HTTP service on", s.ln.Addr().String())
 	s.httpServer.Shutdown(context.Background())
 
 	s.stmtQueue.Close()
@@ -1066,6 +1067,15 @@ func (s *Service) handleStatus(w http.ResponseWriter, r *http.Request) {
 			http.StatusInternalServerError)
 		return
 	}
+
+	key := keyParam(r)
+	b, err = getSubJSON(b, key)
+	if err != nil {
+		http.Error(w, fmt.Sprintf("JSON subkey: %s", err.Error()),
+			http.StatusInternalServerError)
+		return
+	}
+
 	_, err = w.Write(b)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("write: %s", err.Error()),
@@ -1983,6 +1993,44 @@ func isRedirect(req *http.Request) (bool, error) {
 func keyParam(req *http.Request) string {
 	q := req.URL.Query()
 	return strings.TrimSpace(q.Get("key"))
+}
+
+func getSubJSON(jsonBlob []byte, keyString string) (json.RawMessage, error) {
+	if keyString == "" {
+		return jsonBlob, nil
+	}
+
+	keys := strings.Split(keyString, ".")
+	var obj interface{}
+	if err := json.Unmarshal(jsonBlob, &obj); err != nil {
+		return nil, fmt.Errorf("failed to unmarshal json: %w", err)
+	}
+
+	for _, key := range keys {
+		switch val := obj.(type) {
+		case map[string]interface{}:
+			if value, ok := val[key]; ok {
+				obj = value
+			} else {
+				emptyObj := json.RawMessage("{}")
+				return emptyObj, nil
+			}
+		default:
+			// If a value is not a map, marshal and return this value
+			finalObjBytes, err := json.Marshal(obj)
+			if err != nil {
+				return nil, fmt.Errorf("failed to marshal final object: %w", err)
+			}
+			return finalObjBytes, nil
+		}
+	}
+
+	finalObjBytes, err := json.Marshal(obj)
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal final object: %w", err)
+	}
+
+	return finalObjBytes, nil
 }
 
 // timeoutParam returns the value, if any, set for timeout. If not set, it